Used native Algolia types info, nice
